pyproject.toml
[project]
name = "img-squeeze"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "seandlg", email = "16638561+seandlg@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
]

[project.scripts]
img-squeeze = "img_squeeze:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
